requirements.txt
pytube==12.1.0
requests==2.31.0
Pillow==9.4.0
pygame==2.1.2